+++ src/components/search/types.d.ts
@@ -4,4 +4,5 @@ export interface SearchProps extends AriaSearchFieldProps {
   value?: string;
   setValue?: (value: string) => void;
   label?: string;
+  children?: React.ReactNode;
 }